requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
frozenlist==1.6.0
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
multidict==6.4.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pyzmq==26.4.0
requests==2.32.3
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
twilio==9.5.2
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
yarl==1.20.0

# Azure Functions dependencies
azure-functions==1.17.0
